+++ pytorch_forecasting/layers/__init__.py
@@ -1,0 +1,4 @@
+from pytorch_forecasting.layers._dsipts._residual_block_dsipts import ResidualBlock
+from pytorch_forecasting.layers._dsipts._sub_nn import embedding_cat_variables
+
+__all__ = ["ResidualBlock", "embedding_cat_variables"]
