+++ pytorch_forecasting/utils/_utils.py
@@ -16,6 +16,31 @@ import torch.nn.functional as F
 from torch.nn.utils import rnn
 
 
+def beauty_string(message: str, type: str, verbose: bool):
+    import logging
+
+    size = 150
+    if verbose is True:
+        if type == "block":
+            characters = len(message)
+            border = max((100 - characters) // 2 - 5, 0)
+            logging.info("\n")
+            logging.info(f"{'#' * size}")
+            logging.info(f"{'#' * border}{' ' * (size - border * 2)}{'#' * border}")
+            logging.info(f"{message:^{size}}")
+            logging.info(f"{'#' * border}{' ' * (size - border * 2)}{'#' * border}")
+            logging.info(f"{'#' * size}")
+        elif type == "section":
+            logging.info("\n")
+            logging.info(f"{'#' * size}")
+            logging.info(f"{message:^{size}}")
+            logging.info(f"{'#' * size}")
+        elif type == "info":
+            logging.info(f"{message:^{size}}")
+        else:
+            logging.info(message)
+
+
 def integer_histogram(
     data: torch.LongTensor, min: Union[None, int] = None, max: Union[None, int] = None
 ) -> torch.Tensor:
